Tweaked functions to all up to 2^30 or so sites in lattice app.s
git-svn-id: svn+ssh://software-srn.sandia.gov/svn/spparks/trunk@114 685d9ffe-9724-0410-857b-ce06367cce3e

+++ src/memory.cpp
@@ -19,9 +19,10 @@ Memory::Memory(SPK *spk) : SysPtr(spk) {}
    safe malloc 
 ------------------------------------------------------------------------- */
 
-void *Memory::smalloc(int n, const char *name)
+void *Memory::smalloc(long long int n, const char *name)
 {
   if (n == 0) return NULL;
+  //  printf("Attempting to allocate %ld bytes for array %s \n",n,name);
   void *ptr = malloc(n);
   if (ptr == NULL) {
     char str[128];
@@ -46,7 +47,7 @@ void Memory::sfree(void *ptr)
    safe realloc 
 ------------------------------------------------------------------------- */
 
-void *Memory::srealloc(void *ptr, int n, const char *name)
+void *Memory::srealloc(void *ptr, long long int n, const char *name)
 {
   if (n == 0) return NULL;
   ptr = realloc(ptr,n);
